@@ -16,21 +16,20 @@ import com.mattae.simal.modules.base.business.RoleProperties;
 import com.mattae.simal.modules.base.business.RolePropertiesService;
 import com.mattae.simal.modules.base.domain.entities.Module;
 import com.mattae.simal.modules.base.domain.entities.*;
-import com.mattae.simal.modules.base.domain.repositories.MenuRepository;
-import com.mattae.simal.modules.base.domain.repositories.TranslationsRepository;
-import com.mattae.simal.modules.base.domain.repositories.WebComponentRepository;
-import com.mattae.simal.modules.base.domain.repositories.WebRemoteRepository;
+import com.mattae.simal.modules.base.domain.repositories.*;
 import com.mattae.simal.modules.base.yml.ModuleConfig;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
 import org.springframework.util.FileCopyUtils;
 
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
@@ -38,6 +37,7 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class ModuleConfigProcessor {
     private final PermissionService permissionService;
     private final PermissionPropertiesService permissionPropertiesService;
@@ -50,6 +50,7 @@ public class ModuleConfigProcessor {
     private final RoleService roleService;
     private final RoleRepository roleRepository;
     private final TranslationsRepository translationsRepository;
+    private final ConfigurationRepository configurationRepository;
 
     @Transactional
     public void processConfig(ModuleConfig moduleConfig, Module module) {
@@ -138,16 +139,16 @@ public class ModuleConfigProcessor {
                     }).collect(Collectors.toSet());
                 webRemote.setModules(modules);
                 return webRemote;
-            })
-            .collect(Collectors.toList()));
+            }).toList());
 
         module.getWebComponents().addAll(moduleConfig.getWebComponents().stream()
             .map(webComponent -> {
                 webComponent.setModule(module);
                 return webComponent;
-            }).collect(Collectors.toList()));
+            }).toList());
 
         saveTranslations(module, moduleConfig);
+        saveConfigurations(module, moduleConfig);
     }
 
     public void savePermission(List<PermissionProperties> permissionProperties, com.mattae.simal.modules.base.yml.Permission perm,
@@ -167,6 +168,7 @@ public class ModuleConfigProcessor {
     @Transactional
     public void deleteRelationships(Module module) {
         translationsRepository.deleteByModule(module);
+        configurationRepository.deleteByModule(module);
         deleteRolesAndPermissions(module);
         webComponentRepository.deleteAll(webComponentRepository.findByModule(module));
         menuRepository.deleteAll(menuRepository.findByModule(module));
@@ -208,7 +210,56 @@ public class ModuleConfigProcessor {
             });
     }
 
+    @Transactional
+    public void deleteConfigurations(Module module) {
+        configurationRepository.deleteByModule(module);
+    }
+
+    @Transactional
+    public void deleteTranslations(Module module) {
+        translationsRepository.deleteByModule(module);
+    }
+
+    private void saveConfigurations(Module module, ModuleConfig config) {
+        if (config.getConfiguration() != null) {
+            String path = config.getConfiguration().getPath();
+            URL url = module.getClass().getClassLoader().getResource(path);
+            if (url != null) {
+                try {
+                    String data = new String(FileCopyUtils.copyToByteArray(url.openConnection().getInputStream()));
+                    Configuration configuration = new ObjectMapper().readValue(data, Configuration.class);
+                    try (BufferedReader csvReader = new BufferedReader(
+                        new BufferedReader(new InputStreamReader(url.openConnection().getInputStream())))) {
+                        String row;
+                        List<Configuration> configurations = new ArrayList<>();
+                        while ((row = csvReader.readLine()) != null) {
+                            String[] line = row.split(config.getConfiguration().getSeparator());
+                            if (line.length == 4) {
+                                if (List.of("string", "numeric", "bool", "date").contains(line[3])) {
+                                    //Configuration configuration = new Configuration();
+                                    configuration.setCategory(line[0]);
+                                    /*ConfigurationData data = new ConfigurationData();
+                                    data.setKey(line[1]);
+                                    data.setValue(line[2]);
+                                    data.setType(ConfigurationData.ConfigurationType.valueOf(line[3]));
+                                    configuration.setModule(module);*/
+                                    configurations.add(configuration);
+                                }
+                            }
+                        }
+                        configurationRepository.saveAll(configurations);
+                    }
+
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+    }
+
     private void saveTranslations(Module module, ModuleConfig config) {
+        translationsRepository.deleteByModule(module);
+
         if (!config.getTranslations().isEmpty()) {
             config.getTranslations().forEach(tran -> {
                 String path = tran.getPath();
@@ -216,10 +267,10 @@ public class ModuleConfigProcessor {
                 Translation translation = new Translation();
                 translation.setLang(lang);
                 translation.setModule(module);
-                Resource resource = new ClassPathResource(path);
-                if (resource.isFile()) {
+                URL url = module.getClass().getClassLoader().getResource(path);
+                if (url != null) {
                     try {
-                        String data = new String(FileCopyUtils.copyToByteArray(resource.getInputStream()));
+                        String data = new String(FileCopyUtils.copyToByteArray(url.openConnection().getInputStream()));
                         JsonNode node = new ObjectMapper().readTree(data);
                         translation.setData(node);
                         translationsRepository.save(translation);
