@@ -1,16 +1,13 @@
 package com.mattae.simal.modules.base.domain.repositories;
 
+import com.mattae.simal.modules.base.domain.entities.Configuration;
 import com.mattae.simal.modules.base.domain.entities.Module;
-import com.mattae.simal.modules.base.domain.entities.Translation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface TranslationsRepository extends JpaRepository<Translation, Long> {
-    List<Translation> getByLang(String lang);
+public interface ConfigurationRepository extends JpaRepository<Configuration, Long> {
 
     @Modifying
     void deleteByModule(Module module);
